Fixes for Debug build.

+++ lib/FabricDFGBaseInterface.cpp
@@ -486,14 +486,14 @@ void FabricDFGBaseInterface::restoreFromPersistenceData(MString file, MStatus *s
     else
     {
       fseek( file, 0, SEEK_END );
-      int fileSize = ftell( file );
+      long fileSize = ftell( file );
       rewind( file );
 
       char * buffer = (char*) malloc(fileSize + 1);
       buffer[fileSize] = '\0';
 
       size_t readBytes = fread(buffer, 1, fileSize, file);
-      assert(readBytes == fileSize);
+      assert(readBytes == size_t(fileSize));
       (void)readBytes;
 
       fclose(file);
